Migrate CircleBadge to TypeScript
